MM-18241 Fix emoji picker selection and added keyboard tracking on iOS
